Coins now float off of screen
FloatAboveObj no longer floats between two points above an object. This
animation was proving to be too much for the quick pace of the game
Instead coins smoothly lerp into position above the object.
When it is time for the coin to be destroyed it floats off of the screen
(the direction can be specified with the direction variable in the
floatAway method) and is then destroyed
Also apparently changed some stuff in my scene